integracion de los microservicios de evaluaciones y notas

+++ gestionRespaldos/src/main/java/com/example/gestionRespaldos/models/request/ActualizarCopia.java
@@ -10,7 +10,7 @@ public class ActualizarCopia {
     private int id_copia;
 
     @NotBlank
-    private int version;
+    private Double version;
 
     @NotBlank
     private String nombre;
